Exclude selected paths from "git archive"
This commit introduces a work-tree attribute list (see option
"--worktree-attributes" in git-archive(1)) that excludes the ignore
file '.gitignore', the attribute file itself and files whose paths
match '*experimental*'.

Additionally, we also do not distributed the blackoil-sequential or
ad-refactor modules.  The first is not ready for distribution and
the second is a purely internal module for refactoring work.

This is in preparation of MRST 2014b.