Fixes for shear thinning in sequential code.
After running test models, and comparing sequential results
with results from the fully implicit solver, the following
changes are done:
- The shear thinning is NOT included at all in the transport
solver. Including it caused convergence issues, and we did
not see any improvements in the solution.
- If a well control has changed such that the polymer injection
has gone down, we can get a peak in the injector bhp if there
is shear thinning. A special hack has been implemented to fix
this issue. This works at least for the cases we have tested.