Using getRegMap when assigning PVCDO.
The fluid properties assigned by the PVCDO keyword did not use
the getRegMap function, and so did not allow for the
argument 'cellInx'.

Also fixed potential problem of char comparison with colon.
The comparison pvtinx{1}==':' would return true also if pvtinx{1}
has the numerical value of 58. Therefore using strcmp instead.